requirements.txt
semantria-sdk-2x
mysql-connector